+++ gnosis_vpn-lib/src/config/v2.rs
@@ -1,107 +1,19 @@
-use serde::{Deserialize, Deserializer, Serialize};
-use url::Url;
+use serde::{Deserialize, Serialize};
 
 use std::cmp::PartialEq;
 use std::collections::HashMap;
-use std::net::IpAddr;
-use std::net::SocketAddr;
-use std::time::Duration;
 use std::vec::Vec;
 
-use crate::connection::{Destination as ConnDestination, SessionParameters};
-use crate::entry_node::EntryNode;
-use crate::monitor;
+use crate::config::v3;
 use crate::peer_id::PeerId;
-use crate::session;
-use crate::wireguard::config::{Config as WireGuardConfig, ManualMode as WireGuardManualMode};
-
-const MAX_HOPS: u8 = 3;
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Config {
     pub version: u8,
-    hoprd_node: HoprdNode,
-    destinations: Option<HashMap<PeerId, Destination>>,
-    connection: Option<Connection>,
+    hoprd_node: v3::HoprdNode,
+    destinations: Option<HashMap<PeerId, v3::Destination>>,
+    connection: Option<v3::Connection>,
     wireguard: Option<WireGuard>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-struct HoprdNode {
-    endpoint: Url,
-    api_token: String,
-    internal_connection_port: Option<u16>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-struct Destination {
-    meta: Option<HashMap<String, String>>,
-    path: Option<DestinationPath>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-enum DestinationPath {
-    #[serde(alias = "intermediates")]
-    Intermediates(Vec<PeerId>),
-    #[serde(alias = "hops", deserialize_with = "validate_hops")]
-    Hops(u8),
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-struct Connection {
-    listen_host: Option<String>,
-    #[serde(default, with = "humantime_serde::option")]
-    session_timeout: Option<Duration>,
-    bridge: Option<ConnectionProtocol>,
-    wg: Option<ConnectionProtocol>,
-    ping: Option<PingOptions>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-struct ConnectionProtocol {
-    capabilities: Option<Vec<SessionCapability>>,
-    target: Option<SocketAddr>,
-    target_type: Option<SessionTargetType>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-enum SessionCapability {
-    #[serde(alias = "segmentation")]
-    Segmentation,
-    #[serde(alias = "retransmission")]
-    Retransmission,
-}
-
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-enum SessionTargetType {
-    #[default]
-    #[serde(alias = "plain")]
-    Plain,
-    #[serde(alias = "sealed")]
-    Sealed,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-struct ConnectionTarget {
-    bridge: Option<SocketAddr>,
-    wg: Option<SocketAddr>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-struct PingOptions {
-    address: Option<IpAddr>,
-    #[serde(default, with = "humantime_serde::option")]
-    timeout: Option<Duration>,
-    ttl: Option<u32>,
-    seq_count: Option<u16>,
-    #[serde(default, deserialize_with = "validate_ping_interval")]
-    interval: Option<PingInterval>,
-}
-
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-struct PingInterval {
-    min: u8,
-    max: u8,
 }
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
@@ -229,184 +141,25 @@ pub fn wrong_keys(table: &toml::Table) -> Vec<String> {
     wrong_keys
 }
 
-fn validate_hops<'de, D>(deserializer: D) -> Result<u8, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value = u8::deserialize(deserializer)?;
-    if value <= MAX_HOPS {
-        Ok(value)
-    } else {
-        Err(serde::de::Error::custom(format!(
-            "hops must be less than or equal to {MAX_HOPS}"
-        )))
-    }
-}
-
-fn validate_ping_interval<'de, D>(deserializer: D) -> Result<Option<PingInterval>, D::Error>
-where
-    D: Deserializer<'de>,
-{
-    let value: Option<PingInterval> = Option::deserialize(deserializer)?;
-    match value {
-        Some(interval) => {
-            if interval.min < interval.max {
-                Ok(Some(interval))
-            } else {
-                Err(serde::de::Error::custom("min must be less than max"))
-            }
-        }
-        None => Ok(None),
-    }
-}
-
-impl From<SessionCapability> for session::Capability {
-    fn from(val: SessionCapability) -> Self {
-        match val {
-            SessionCapability::Segmentation => session::Capability::Segmentation,
-            SessionCapability::Retransmission => session::Capability::Retransmission,
+impl From<&Config> for v3::Config {
+    fn from(config: &Config) -> Self {
+        v3::Config {
+            version: config.version,
+            hoprd_node: config.hoprd_node.clone(),
+            destinations: config.destinations.clone(),
+            connection: config.connection.clone(),
+            wireguard: config.wireguard.as_ref().map(|wg| wg.into()),
         }
     }
 }
 
-impl Connection {
-    pub fn default_bridge_capabilities() -> Vec<SessionCapability> {
-        vec![SessionCapability::Segmentation, SessionCapability::Retransmission]
-    }
-
-    pub fn default_wg_capabilities() -> Vec<SessionCapability> {
-        vec![SessionCapability::Segmentation]
-    }
-
-    pub fn default_bridge_target() -> SocketAddr {
-        SocketAddr::from(([172, 30, 0, 1], 8000))
-    }
-
-    pub fn default_wg_target() -> SocketAddr {
-        SocketAddr::from(([172, 30, 0, 1], 51820))
-    }
-
-    pub fn default_listen_host() -> String {
-        ":1422".to_string()
-    }
-
-    pub fn default_session_timeout() -> Duration {
-        Duration::from_secs(15)
-    }
-
-    pub fn default_ping_interval() -> PingInterval {
-        PingInterval { min: 5, max: 10 }
-    }
-}
-
-impl Config {
-    pub fn entry_node(&self) -> EntryNode {
-        let internal_connection_port = self.hoprd_node.internal_connection_port.map(|p| format!(":{p}"));
-        let listen_host = self
-            .connection
-            .as_ref()
-            .and_then(|c| c.listen_host.clone())
-            .or(internal_connection_port)
-            .unwrap_or(Connection::default_listen_host());
-        let session_timeout = self
-            .connection
-            .as_ref()
-            .and_then(|c| c.session_timeout)
-            .unwrap_or(Connection::default_session_timeout());
-        EntryNode::new(
-            &self.hoprd_node.endpoint,
-            &self.hoprd_node.api_token,
-            &listen_host,
-            &session_timeout,
-        )
-    }
-
-    pub fn destinations(&self) -> HashMap<PeerId, ConnDestination> {
-        let config_dests = self.destinations.clone().unwrap_or_default();
-        let connection = self.connection.as_ref();
-        config_dests
-            .iter()
-            .map(|(k, v)| {
-                let path = match v.path.clone() {
-                    Some(DestinationPath::Intermediates(p)) => session::Path::IntermediatePath(p),
-                    Some(DestinationPath::Hops(h)) => session::Path::Hops(h),
-                    None => session::Path::default(),
-                };
-
-                let bridge_caps = connection
-                    .and_then(|c| c.bridge.as_ref())
-                    .and_then(|b| b.capabilities.clone())
-                    .unwrap_or(Connection::default_bridge_capabilities())
-                    .iter()
-                    .map(|cap| <SessionCapability as Into<session::Capability>>::into(cap.clone()))
-                    .collect::<Vec<session::Capability>>();
-                let bridge_target_socket = connection
-                    .and_then(|c| c.bridge.as_ref())
-                    .and_then(|b| b.target)
-                    .unwrap_or(Connection::default_bridge_target());
-                let bridge_target_type = connection
-                    .and_then(|c| c.bridge.as_ref())
-                    .and_then(|b| b.target_type.clone())
-                    .unwrap_or_default();
-                let bridge_target = match bridge_target_type {
-                    SessionTargetType::Plain => session::Target::Plain(bridge_target_socket),
-                    SessionTargetType::Sealed => session::Target::Sealed(bridge_target_socket),
-                };
-                let params_bridge = SessionParameters::new(&bridge_target, &bridge_caps);
-
-                let wg_caps = connection
-                    .and_then(|c| c.wg.as_ref())
-                    .and_then(|w| w.capabilities.clone())
-                    .unwrap_or(Connection::default_wg_capabilities())
-                    .iter()
-                    .map(|cap| <SessionCapability as Into<session::Capability>>::into(cap.clone()))
-                    .collect::<Vec<session::Capability>>();
-                let wg_target_socket = connection
-                    .and_then(|c| c.wg.as_ref())
-                    .and_then(|w| w.target)
-                    .unwrap_or(Connection::default_wg_target());
-                let wg_target_type = connection
-                    .and_then(|c| c.wg.as_ref())
-                    .and_then(|w| w.target_type.clone())
-                    .unwrap_or_default();
-                let wg_target = match wg_target_type {
-                    SessionTargetType::Plain => session::Target::Plain(wg_target_socket),
-                    SessionTargetType::Sealed => session::Target::Sealed(wg_target_socket),
-                };
-                let params_wg = SessionParameters::new(&wg_target, &wg_caps);
-                let meta = v.meta.clone().unwrap_or_default();
-
-                let interval = connection
-                    .and_then(|c| c.ping.as_ref())
-                    .and_then(|p| p.interval.clone())
-                    .unwrap_or(Connection::default_ping_interval());
-
-                let def_opts = monitor::PingOptions::default();
-                let opts = connection
-                    .and_then(|c| c.ping.as_ref())
-                    .map(|p| monitor::PingOptions {
-                        address: p.address.unwrap_or(def_opts.address),
-                        timeout: p.timeout.unwrap_or(def_opts.timeout),
-                        ttl: p.ttl.unwrap_or(def_opts.ttl),
-                        seq_count: p.seq_count.unwrap_or(def_opts.seq_count),
-                    })
-                    .unwrap_or(def_opts);
-                let ping_range = interval.min..interval.max;
-                let dest = ConnDestination::new(*k, path, meta, params_bridge, params_wg, ping_range, opts);
-                (*k, dest)
-            })
-            .collect()
-    }
-
-    pub fn wireguard(&self) -> WireGuardConfig {
-        let listen_port = self.wireguard.as_ref().and_then(|wg| wg.listen_port);
-        let allowed_ips = self.wireguard.as_ref().and_then(|wg| wg.allowed_ips.clone());
-        let manual_mode = self
-            .wireguard
-            .as_ref()
-            .and_then(|wg| wg.manual_mode.as_ref())
-            .map(|wgm| WireGuardManualMode::new(wgm.public_key.clone()));
-        WireGuardConfig::new(listen_port, allowed_ips, manual_mode)
+impl From<&WireGuard> for v3::WireGuard {
+    fn from(wg: &WireGuard) -> Self {
+        v3::WireGuard {
+            listen_port: wg.listen_port,
+            allowed_ips: wg.allowed_ips.clone(),
+            force_private_key: None,
+        }
     }
 }
 
@@ -421,10 +174,6 @@ version = 2
 [hoprd_node]
 endpoint = "http://127.0.0.1:3001"
 api_token = "1234567890"
-
-[connection.ping]
-address = "10.128.0.1"
-
 "#####;
         toml::from_str::<Config>(config).expect("Failed to parse minimal config");
     }
@@ -436,6 +185,9 @@ version = 2
 [hoprd_node]
 endpoint = "http://127.0.0.1:3001"
 api_token = "1234567890"
+
+[connection.ping]
+address = "10.128.0.1"
 "#####;
         toml::from_str::<Config>(config).expect("Failed to parse minimal ping");
     }
